% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{eval_mnist_accuracy}
\alias{eval_mnist_accuracy}
\title{Evaluate MNIST Accuracy}
\usage{
eval_mnist_accuracy(model, ds, device = "cpu", batch_size = 1000)
}
\arguments{
\item{model}{The model to evaluate.}

\item{ds}{The dataset to evaluate on.}

\item{device}{The device to use ("cpu").}

\item{batch_size}{Batch size.}
}
\value{
The accuracy (0-1).
}
\description{
Computes accuracy on a dataset.
}
