% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedavg.R
\name{fedavg}
\alias{fedavg}
\title{Federated Averaging Aggregation}
\usage{
fedavg(params_list, weights)
}
\arguments{
\item{params_list}{A list of flattened parameter vectors from clients.}

\item{weights}{A numeric vector of weights (usually number of samples).}
}
\value{
A single flattened parameter vector.
}
\description{
Aggregates client updates using weighted averaging.
}
