% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{flatten_params}
\alias{flatten_params}
\title{Flatten Model Parameters}
\usage{
flatten_params(model)
}
\arguments{
\item{model}{A torch module.}
}
\value{
A numeric vector containing all parameters.
}
\description{
Flattens all parameters of a torch module into a single numeric vector.
}
