% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_set_params}
\alias{get_set_params}
\title{Get or Set Model Parameters}
\usage{
get_set_params(model, params = NULL)
}
\arguments{
\item{model}{A torch module.}

\item{params}{Optional. If provided, updates the model parameters.}
}
\value{
If params is NULL, returns the flattened parameters. Otherwise returns the updated model.
}
\description{
Get or Set Model Parameters
}
