% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitions.R
\name{iid_split}
\alias{iid_split}
\title{IID Split}
\usage{
iid_split(n_items, K, seed = NULL)
}
\arguments{
\item{n_items}{Total number of items.}

\item{K}{Number of clients.}

\item{seed}{Random seed.}
}
\value{
A list of K integer vectors (indices).
}
\description{
Randomly partitions data into K clients.
}
