% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loaders.R
\name{mnist_ds}
\alias{mnist_ds}
\title{MNIST Dataset}
\usage{
mnist_ds(root = "data", train = TRUE, download = TRUE)
}
\arguments{
\item{root}{Directory to store data.}

\item{train}{Boolean, whether to load training set.}

\item{download}{Boolean, whether to download if missing.}
}
\value{
A torch dataset.
}
\description{
Wrapper for torchvision::mnist_dataset with required transforms.
Transform matches Flower reference implementation.
}
\references{
Mirrored from reference/fedavg_mnist_flwr/dataset.py:69-71
}
