% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loaders.R
\name{mnist_labels}
\alias{mnist_labels}
\title{Get MNIST Labels}
\usage{
mnist_labels(ds, batch_size = 2000)
}
\arguments{
\item{ds}{The MNIST dataset.}

\item{batch_size}{Batch size for loading.}
}
\value{
A numeric vector of labels.
}
\description{
Extracts labels from the dataset for partitioning.
}
