% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_mlp.R
\name{mnist_mlp}
\alias{mnist_mlp}
\title{MNIST MLP Model (2NN)}
\usage{
mnist_mlp()
}
\value{
A torch module.
}
\description{
Creates a 2-layer fully connected neural network for MNIST (784 -> 200 -> 200 -> 10).
}
