% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitions.R
\name{sample_clients}
\alias{sample_clients}
\title{Sample Clients}
\usage{
sample_clients(K, C, seed = NULL)
}
\arguments{
\item{K}{Total clients.}

\item{C}{Fraction of clients to select.}

\item{seed}{Random seed.}
}
\value{
Sorted integer vector of selected client indices.
}
\description{
Selects a fraction of clients.
}
